Add PrintAvailableClasses method to ClassActions
A new method `PrintAvailableClasses` has been added to the `ClassActions` class within the `GymManagement.Console.ConsoleUtilities` namespace. This method uses a `UnitOfWork` to retrieve a list of available classes asynchronously. If no classes are available, it prints a message indicating that there are no available classes. If there are available classes, it prints the details of each class, including the class ID, name, description, scheduled date, duration, trainer ID, and availability status.