Refactor async methods to return boolean values
Changed return type of HandleAdminActionsAsync and HandleTrainerActionsAsync from Task to Task<bool> in AdminActions.cs and TrainerActions.cs respectively. Added return false statements at the end of these methods. Updated RoleMenus.cs to handle the boolean return values, exiting the loop if the methods return true or the input is "0".